diverse oppgraderinger

+++ src/main/kotlin/no/nav/syfo/service/RedisService.kt
@@ -7,8 +7,8 @@ import java.security.MessageDigest
 import java.util.concurrent.TimeUnit
 
 fun updateRedis(jedis: Jedis, ediLoggId: String, sha256String: String) {
-    jedis.setex(ediLoggId, TimeUnit.DAYS.toSeconds(7).toInt(), ediLoggId)
-    jedis.setex(sha256String, TimeUnit.DAYS.toSeconds(7).toInt(), ediLoggId)
+    jedis.setex(ediLoggId, TimeUnit.DAYS.toSeconds(7), ediLoggId)
+    jedis.setex(sha256String, TimeUnit.DAYS.toSeconds(7), ediLoggId)
 }
 
 fun sha256hashstring(helseOpplysningerArbeidsuforhet: HelseOpplysningerArbeidsuforhet): String =
